fix one failing test
